Create base CollectionViewCell to avoid file pollution
Allow its subclasses to be self-contained
